hive/users/serializers.py
from rest_framework import serializers
from hive.users.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['first_name', 'last_name', 'username', 'email', 'last_login']
