pyproject.toml
[tool.poetry]
name = "hive-exch"
version = "0.1.0"
description = ""
authors = ["dmkael <tanmaar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
dj-database-url = "^2.2.0"
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
django-tz-detect = "^0.5.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
exchangelib = "^5.4.2"
pyasn1 = "^0.6.0"
pycryptodomex = "^3.20.0"
gssapi = "^1.8.3"
pyopenssl = "^24.1.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
